提问的输入金额参数和私圈的咨询费用是否相等校验

+++ resource/service/src/main/java/com/yryz/quanhu/resource/questionsAnswers/service/impl/QuestionServiceImpl.java
@@ -144,7 +144,7 @@ public class QuestionServiceImpl implements QuestionService {
          * 校验提问金额是否合法
          */
         Integer consultingFee = coterieInfo.getConsultingFee() == null ? 0 : coterieInfo.getConsultingFee();
-        if (question.getChargeAmount() == null || question.getChargeAmount() != Long.valueOf(consultingFee)) {
+        if (question.getChargeAmount() == null || question.getChargeAmount().longValue() != Long.valueOf(consultingFee).longValue()) {
             throw new QuanhuException(ExceptionEnum.AMOUNT_PARAM_ERROR);
         }
         //保存提问费用
